Fix wrong behavior when using mumble::default without mysql
The changes in ced66ce71b656eb1f583505498e358a973fcc1ce made it impossible to setup murmur without mysql because attributes/mysql.rb is always executed by (at least my version [10.16] of) Chef and not only when a recipe with matching name is run. This seems to be a common misunderstanding of Chef attribute handling although I'm not able to find a reference in the public Chef doc - at least in my tests this holds true.

(Even when setting the node['mumble']['mysql'] property to nil there is an error because set_unless['mumble']['mysql']['password'] has then trouble figuring out the if there is already a value set...)